Pick up new API from External Logging API
